fix: now can be built for Windows, but will panic at runtime
